# docs/swagger.yaml
basePath: /
definitions:
  go_auth_internal_apierror.Body:
    properties:
      error:
        $ref: '#/definitions/go_auth_internal_apierror.Detail'
    type: object
  go_auth_internal_apierror.Detail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  go_auth_internal_application.Application:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      uuid:
        type: string
    type: object
...
  go_auth_internal_role.Role:
    properties:
      active:
        type: boolean
      application_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      uuid:
        type: string
    type: object
  go_auth_internal_user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
      uuid:
        type: string
    type: object
  internal_access.AssignRequest:
    properties:
      application_id:
        type: integer
      profile_id:
        type: integer
      role_id:
        type: integer
    required:
    - application_id
    - role_id
    type: object
  internal_access.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_access.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_access.UserInfo'
    type: object
  internal_access.UserApplicationAccess:
    properties:
      application:
        $ref: '#/definitions/go_auth_internal_application.Application'
      permissions:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
    type: object
  internal_access.UserApplicationRole:
    properties:
      application_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      profile_id:
        type: integer
      role_id:
        type: integer
      user_id:
        type: integer
    type: object
  internal_access.UserInfo:
    properties:
      applications:
        items:
          $ref: '#/definitions/internal_access.UserApplicationAccess'
        type: array
      email:
        type: string
      id:
        type: integer
      username:
        type: string
      uuid:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API de autenticação e autorização em Go com JWT, suportando múltiplas
    aplicações
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Auth API
  version: "1.0"
paths:
  /api/admin/application-roles/users/{userId}/applications/{applicationId}/roles:
    get:
      consumes:
      - application/json
      description: Retrieve all roles assigned to a user within a specific application
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go_auth_internal_role.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get user roles in application
      tags:
      - Admin - User-Application-Role Relationships
...
  /api/admin/applications/{id}/users:
    get:
      consumes:
      - application/json
      description: Retrieve all users assigned to a specific application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go_auth_internal_user.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get application users
      tags:
      - Admin - User-Application-Role Relationships
...
  /api/admin/users/{id}/application-roles:
    get:
      consumes:
      - application/json
      description: Retrieve all application-role assignments for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_access.UserApplicationRole'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get user application roles
      tags:
      - Admin - User-Application-Role Relationships
    post:
      consumes:
      - application/json
      description: Assign a specific role to a user within an application context
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/internal_access.AssignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Assign role to user in application
      tags:
      - Admin - User-Application-Role Relationships
  /api/admin/users/{id}/application-roles/{applicationId}/{roleId}:
    delete:
      consumes:
      - application/json
      description: Remove a specific role from a user within an application context
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: integer
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Remove role from user in application
      tags:
      - Admin - User-Application-Role Relationships
  /api/admin/users/{id}/applications:
    get:
      consumes:
      - application/json
      description: Retrieve all applications assigned to a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go_auth_internal_application.Application'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get user applications
      tags:
      - Admin - User-Application-Role Relationships
  /api/admin/users/{id}/applications/{applicationId}:
    delete:
      consumes:
      - application/json
      description: Revoke every role assignment a user holds within an application
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Remove all roles from a user in an application
      tags:
      - Admin - User-Application-Role Relationships
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT token with
        user info
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_access.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_access.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      summary: User login
      tags:
      - Authentication
  /api/me:
    get:
      consumes:
      - application/json
      description: Get current authenticated user information including roles and
        permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_access.UserInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go_auth_internal_apierror.Body'
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - Authentication
...
securityDefinitions:
  Bearer:
    description: Digite "Bearer " seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"